\encoding{utf8} 
\name{basicInference-internal}
\alias{basicInference-internal}
\alias{summary.basicInference}
\alias{print.summary.basicInference}
\alias{plot.basicInference}

\title{Internal basicInference functions}
\description{Internal basicInference functions}
\usage{
\method{summary}{basicInference}(object, ...)
\method{print.summary}{basicInference}(x, ...)
\method{plot}{basicInference}(x, ...)
}

\details{These functions are not to be called by the user}
\seealso{
\code{\link{t_test}}
}
\keyword{internal}
