% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Conducts a t test and gets the sample means, test statistic and p-value}
\usage{
t_test(sample1, sample2, cl = 0.95)
}
\arguments{
\item{sample1}{Vector with observations from sample1.}

\item{sample2}{Vector with observations from sample2.}

\item{cl}{Confidence level.}
}
\value{
A vector with sample means, test statistic and p-value.
}
\description{
Conducts a t test and gets the sample means, test statistic and p-value
}
\examples{
s1 <- rnorm(500)
s2 <- rnorm(500) 
t_test(s1, s2)
}
